type casting

+++ knowrob_learning/knowrob_learning/src/main/java/org/knowrob/reinforcement/KitchenEnvironment.java
@@ -818,8 +818,8 @@ public class KitchenEnvironment extends AbstractNodeMain implements Environment
 			else
 			{
 				Pr2GripperCommand gripper_cmd = gripper_publisher.newMessage();
-				gripper_cmd.setPosition(0.0);
-				gripper_cmd.setMaxEffort(100.0);
+				gripper_cmd.setPosition((float)0.0);
+				gripper_cmd.setMaxEffort((float)100.0);
 				gripper_publisher.publish(gripper_cmd);
 			
 				try
